moe_fp8_cv_weight_ag round 3: in-body e4m3 gather -> ONE direct bf16 weight-grad RS per weight
Root cause of the round-2 +0.35 RS regression: with the weight entering the main shard_map
REPLICATED (mini-shard_map pregather), the shard_map transpose ARs the weight grad over fsdp and
the custom_vjp reshard slices it -- XLA failed to fuse AR+slice into a reduce-scatter (baseline
fuses it: one 962MB bf16 RS @ 235 GB/s), leaving 2 slow RSs (283+262MB @ 22-38 GB/s) + extra AR.

Fix (provably-single-reduction by construction, not fusion luck):
- gather_weights fp8 path: quantize the STORAGE-SHARDED param via _ste_quant only (no gather);
  return the SHARDED e4m3 qvalue QArray. _make_cv_quant_gather deleted.
- sparse_matmul: split the QArrays BEFORE the shard_map decorator and OVERRIDE the weight in_specs
  to the STORAGE sharding (fsdp on embed); gather INSIDE the body top, once, reused across chunks:
  jax.lax.all_gather(qv,'fsdp',axis=1/2,tiled=True). Autodiff transposes that gather into ONE
  direct psum_scatter of the bf16 weight grad to storage sharding; the boundary adds no reduction
  (input varying, not replicated) -> no x128 overcount, no AR+slice.
- ops._gmm_bwd: cast the drhs qvalue cotangent to BF16 (gradient wire dtype; uncast f32 doubled
  the RS bytes). Never e4m3.
- moe_fp8_cv_weight_ag_tags DEPRECATED/ignored (NaN'd + dominated on cluster round 2; nothing to
  tag with the in-body gather).

Numerical receipt (test_cv_fp8_grad_v2.py, 8-way CPU mesh, data-parallel per-shard tokens):
d w cos=+0.99965 ratio=0.9971; d x cos=+0.99951 ratio=0.9972.
AOT gate: routed e4m3 AGs intact (f8e4m3fn[32,7168,2048]x4 + [32,2048,7168]x2); weight-grad RS =
ONE bf16 RS per weight (w0/w1 as bf16[56,65536]x2 reshaped, wo as bf16[32,2048,56], scope
transpose(jvp)/checkpoint/shard_map/reduce_scatter); bwd tiny collectives none; bwd all-reduce 22;
temp 44.79GB; Finished clean.

+++ src/maxtext/kernels/megablox/ops.py
@@ -553,12 +553,14 @@ def _gmm_bwd(
   # TODO(tgale, enriqueps, apaske): Fuse this transposition into the tgmm.
   drhs = drhs.swapaxes(1, 2) if transpose_rhs else drhs
   if _orig_rhs_qarray is not None:
-    # The primal rhs entered as a QArray (moe_fp8_boundary_qag), so the VJP cotangent must match its
-    # pytree. Keep the qvalue-leaf cotangent (the weight gradient) in its NATURAL bf16/f32 dtype --
-    # NEVER cast it to e4m3 (that overflowed -> NaN and blew up the weight reduce-scatter). The scale
-    # is forward-only (STE on the quant), so its cotangent is a per-channel zero.
+    # The primal rhs entered as a QArray (moe_fp8_cv_weight_ag / moe_fp8_boundary_qag), so the VJP
+    # cotangent must match its pytree. Cast the qvalue-leaf cotangent (the weight gradient) to BF16:
+    # it is the gradient WIRE dtype -- the in-body all_gather transpose reduce-scatters it to storage
+    # sharding, and an uncast f32 drhs doubles that RS's bytes vs the baseline's bf16 weight-grad RS.
+    # NEVER cast to e4m3 (overflow -> NaN, the round-1 bug). The scale is forward-only (STE on the
+    # quant), so its cotangent is a per-channel zero.
     drhs = qpl.QArray(
-        qvalue=drhs,
+        qvalue=drhs.astype(jnp.bfloat16),
         scale=jnp.zeros_like(_orig_rhs_qarray.scale),
         zero_point=None,
         qtype=_orig_rhs_qarray.qtype,
